+++ src/automatic_differentiation.rs
@@ -6,3 +6,4 @@ pub(crate) mod directional_derivative;
 pub(crate) mod dual;
 pub(crate) mod dual_vector;
 pub(crate) mod gradient;
+pub(crate) mod partial_derivative;